test: 拆分 Respond 记忆与会话测试
